backend/requirements.txt
fastapi
uvicorn
python-multipart
mediapipe
opencv-python
numpy
pillow
